added explicit sign convertions to most of geometry and moveresize function to try and nail down potential signed/unsigned bugs

+++ libAfterStep/canvas.h
@@ -59,11 +59,11 @@ typedef struct ASCanvas
 #define ASWidget ASCanvas
 #define AS_WIDGET_SCREEN(c) (ASDefaultScr)
 #define AS_WIDGET_WINDOW(c) ((c)?(c)->w:None)
-#define AS_WIDGET_X(c)      ((c)?(c)->root_x:0)
-#define AS_WIDGET_Y(c)      ((c)?(c)->root_y:0)
-#define AS_WIDGET_WIDTH(c)  ((c)?(c)->width:0)
-#define AS_WIDGET_HEIGHT(c) ((c)?(c)->height:0)
-#define AS_WIDGET_BW(c) 	((c)?(c)->bw:0)
+#define AS_WIDGET_X(c)      ((c)?(int)(c)->root_x:0)
+#define AS_WIDGET_Y(c)      ((c)?(int)(c)->root_y:0)
+#define AS_WIDGET_WIDTH(c)  ((c)?(int)(c)->width:0)
+#define AS_WIDGET_HEIGHT(c) ((c)?(int)(c)->height:0)
+#define AS_WIDGET_BW(c) 	((c)?(int)(c)->bw:0)
 #define AS_WIDGET_FEEL(c)   (&(ASDefaultScr->Feel))
 #define AS_WIDGET_LOOK(c)   (&(ASDefaultScr->Look))
